Persisted chargeprices
