+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/ChargeCommands/Handlers/ChargeCommandBundleHandler.cs
@@ -13,9 +13,10 @@
 // limitations under the License.
 
 using System.Threading.Tasks;
+using GreenEnergyHub.Charges.Application.Charges.Handlers;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands;
 
-namespace GreenEnergyHub.Charges.Application.Charges.Handlers
+namespace GreenEnergyHub.Charges.Application.ChargeCommands.Handlers
 {
     public class ChargeCommandBundleHandler : IChargesBundleHandler
     {
