+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/ChargeCommands/Handlers/ChargeCommandReceivedEventHandler.cs
@@ -14,7 +14,8 @@
 
 using System;
 using System.Threading.Tasks;
-using GreenEnergyHub.Charges.Application.Charges.Acknowledgement;
+using GreenEnergyHub.Charges.Application.ChargeCommands.Acknowledgement;
+using GreenEnergyHub.Charges.Application.Charges.Handlers;
 using GreenEnergyHub.Charges.Application.Messaging;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommandAcceptedEvents;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommandReceivedEvents;
@@ -22,24 +23,27 @@ using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands;
 using GreenEnergyHub.Charges.Domain.Dtos.Validation;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
 
-namespace GreenEnergyHub.Charges.Application.Charges.Handlers
+namespace GreenEnergyHub.Charges.Application.ChargeCommands.Handlers
 {
-    public class ChargeHandler : IChargeHandler
+    public class ChargeCommandReceivedEventHandler : IChargeCommandReceivedEventHandler
     {
-        private readonly IChargeCommandReceivedEventHandler _chargeCommandReceivedEventHandler;
+        private readonly IChargeEventHandler _chargeEventHandler;
+        private readonly IChargePriceEventHandler _chargePriceEventHandler;
         private readonly IMessageDispatcher<ChargeCommandAcceptedEvent> _acceptedMessageDispatcher;
         private readonly IChargeCommandAcceptedEventFactory _chargeCommandAcceptedEventFactory;
         private readonly IDocumentValidator<ChargeCommand> _documentValidator;
         private readonly IChargeCommandReceiptService _chargeCommandReceiptService;
 
-        public ChargeHandler(
-            IChargeCommandReceivedEventHandler chargeCommandReceivedEventHandler,
+        public ChargeCommandReceivedEventHandler(
+            IChargeEventHandler chargeEventHandler,
+            IChargePriceEventHandler chargePriceEventHandler,
             IMessageDispatcher<ChargeCommandAcceptedEvent> acceptedMessageDispatcher,
             IChargeCommandAcceptedEventFactory chargeCommandAcceptedEventFactory,
             IDocumentValidator<ChargeCommand> documentValidator,
             IChargeCommandReceiptService chargeCommandReceiptService)
         {
-            _chargeCommandReceivedEventHandler = chargeCommandReceivedEventHandler;
+            _chargeEventHandler = chargeEventHandler;
+            _chargePriceEventHandler = chargePriceEventHandler;
             _acceptedMessageDispatcher = acceptedMessageDispatcher;
             _chargeCommandAcceptedEventFactory = chargeCommandAcceptedEventFactory;
             _documentValidator = documentValidator;
@@ -59,13 +63,10 @@ namespace GreenEnergyHub.Charges.Application.Charges.Handlers
             switch (commandReceivedEvent.Command.Document.BusinessReasonCode)
             {
                 case BusinessReasonCode.UpdateChargePrices:
-                    var chargeCommandAcceptedEvent =
-                        _chargeCommandAcceptedEventFactory.CreateEvent(commandReceivedEvent.Command);
-                    await _acceptedMessageDispatcher.DispatchAsync(chargeCommandAcceptedEvent)
-                        .ConfigureAwait(false);
+                    await _chargePriceEventHandler.HandleAsync(commandReceivedEvent).ConfigureAwait(false);
                     break;
                 case BusinessReasonCode.UpdateChargeInformation:
-                    await _chargeCommandReceivedEventHandler.HandleAsync(commandReceivedEvent).ConfigureAwait(false);
+                    await _chargeEventHandler.HandleAsync(commandReceivedEvent).ConfigureAwait(false);
                     break;
                 default:
                     throw new ArgumentOutOfRangeException(
