+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/ChargeCommands/Handlers/ChargeIntegrationEventsPublisher.cs
@@ -16,9 +16,10 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using GreenEnergyHub.Charges.Application.Charges.Acknowledgement;
+using GreenEnergyHub.Charges.Application.Charges.Handlers;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommandAcceptedEvents;
 
-namespace GreenEnergyHub.Charges.Application.Charges.Handlers
+namespace GreenEnergyHub.Charges.Application.ChargeCommands.Handlers
 {
     public class ChargeIntegrationEventsPublisher : IChargeIntegrationEventsPublisher
     {
