+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Charges/ChargeCommandReceiverEndpoint.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System.Threading.Tasks;
+using GreenEnergyHub.Charges.Application.ChargeCommands.Handlers;
 using GreenEnergyHub.Charges.Application.Charges.Handlers;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommandReceivedEvents;
 using GreenEnergyHub.Charges.FunctionHost.Common;
@@ -24,14 +25,14 @@ namespace GreenEnergyHub.Charges.FunctionHost.Charges
     public class ChargeCommandReceiverEndpoint
     {
         public const string FunctionName = nameof(ChargeCommandReceiverEndpoint);
-        private readonly IChargeHandler _chargeHandler;
+        private readonly IChargeCommandReceivedEventHandler _chargeCommandReceivedEventCommandReceivedEventHandler;
         private readonly JsonMessageDeserializer<ChargeCommandReceivedEvent> _deserializer;
 
         public ChargeCommandReceiverEndpoint(
-            IChargeHandler chargeHandler,
+            IChargeCommandReceivedEventHandler chargeCommandReceivedEventCommandReceivedEventHandler,
             JsonMessageDeserializer<ChargeCommandReceivedEvent> deserializer)
         {
-            _chargeHandler = chargeHandler;
+            _chargeCommandReceivedEventCommandReceivedEventHandler = chargeCommandReceivedEventCommandReceivedEventHandler;
             _deserializer = deserializer;
         }
 
@@ -44,7 +45,7 @@ namespace GreenEnergyHub.Charges.FunctionHost.Charges
             byte[] message)
         {
             var receivedEvent = (ChargeCommandReceivedEvent)await _deserializer.FromBytesAsync(message).ConfigureAwait(false);
-            await _chargeHandler.HandleAsync(receivedEvent).ConfigureAwait(false);
+            await _chargeCommandReceivedEventCommandReceivedEventHandler.HandleAsync(receivedEvent).ConfigureAwait(false);
         }
     }
 }
